+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/plannerLogin/Planner.java
@@ -1,24 +1,26 @@
-package com.mysite.weddingyou_backend.user.login;
+package com.mysite.weddingyou_backend.plannerLogin;
 
 import java.time.LocalDateTime;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity // 물리적인 테이블을 생성
 @Setter
 @Getter
-@Table(name = "user") //UserEntity클래스를 사용해서 user라는 테이블이 만들어짐
-public class User {
+@Table(name = "planner") //PlannerEntity클래스를 사용해서 planner라는 테이블이 만들어짐
+public class Planner {
 	@Id // pk 지정
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "user_id")
-	private int userId;
+	@Column(name = "planner_id")
+	private int plannerId;
 
 	@Column(name = "name", nullable = false)
 	private String name;
@@ -32,11 +34,14 @@ public class User {
 	@Column(name = "phone_number", nullable = false)
 	private String phoneNum;
 
-	@Column(name = "user_img", nullable = true)
-	private byte[] userImg;
+	@Column(name = "planner_img", nullable = true)
+	private byte[] plannerImg;
+	
+	@Column(name = "planner_career_years", nullable = false)
+	private String plannerCareerYears;
 
-	@Column(name = "user_join_date", columnDefinition = "datetime default current_timestamp")
-	private LocalDateTime userJoinDate;
+	@Column(name = "planner_join_date", columnDefinition = "datetime default current_timestamp")
+	private LocalDateTime plannerJoinDate;
 	
 
 }
